restructured the type system so that the expression types are static
I wanted the different expression types to be static rather then just containing
an enum and dealing with it at run time.
I decided to make some internal types and use phantom data.
If I have done this right I people outside of the expression mod will not be
able to make new expression types.
Normally I would want to use a enum to capture the limited scope but it looks
like rust dose not yet support variants as types for genetics.